meson: create unit symlinks in the vendor directory
During the v242 cycle, a change was merged [1] which stopped creating
enablement symlinks for various services, including systemd-networkd.
The main reason described in the commit message was that this
auto-provisioning undoes configuration for those end users who chose to
disable some of these services on their systems. For them, running
`ninja install` caused these services to become enabled again, which was
not expected. Running `systemctl preset-all` is documented to be a
recommended step prior to the first boot.

This change has implications for cross-compilations. A build machine
cannot execute binaries (such as `sysrepoctl`) which were just built for
the target. There are workarounds for this, including QEMU, but they are
needlesly complex. Creating these symlinks by hand works, but it's yet
another manual step which can get out of sync from the actual code.

Systemd has a concept of vendor presets, bits which are intended to
represent default configuration "from the factory". These settings are
in no way affected by the end-user configuration. This patch simply
takes all services which were, prior to 01d2041e41, enabled by symlinks
in `/etc/systemd/system/`, and puts these into an appropriate `*.wants`
below the `/lib/systemd/system/`.  This means that the usual way of
overriding and masking stuff via `/etc` is available to express a user's
configuration. As before, stuff below `/lib` represents the default
system state, and people are free to override this "policy" via a
configuration below `/etc`.

There's one change for people who already grew accustomed to the v242's
behavior. They will now have to mask each corresponding unit (because
this "enablement" is not an "enablement" as in `systemctl is-enabled`).

Closes #12767, #12769

[1] https://github.com/systemd/systemd/commit/01d2041e41f4886a6dff16a53a950ae8d5b66bc7

Signed-off-by: Jan Kundrát <jan.kundrat@cesnet.cz>