travis: run Coverity after Fuzzit-Fuzzing
Coverity is unpredictable and, according to a notification I received
yeserday, it will be upgraded on June 17. During the upgrade
it might be offline for 3 days, af far as I understand. Anyway, Travis
stops as soon as a stage fails so it makes sense to put stages that
are likely to fail at the end so that the others have a chance to
do what they are supposed to do.

https://community.synopsys.com/s/topic/0TO2H0000001CN7WAM/coverity-scan-status