device: rework the logic use to trust device enumeration
Rather than assuming that the device enumeration can be trusted based on the
manager state (reloading vs reexecuting), we now simply look at the udev DB
itself - ie if no single devices couldn't have been enumerated then we assume
that the DB is not yet initialized [1] and thereby don't trust the enumeration
phase.

In this case we restore the serialized state and rely on events retriggering to
update the view on device PID1 had before it reexecuted.

[1] Maybe there's a better way to detect this state.

Fixes: #12953