coredump: retrieve UID and GID metadata from /proc
The kernel passes their value (through %u and %g) as seen in the host's user
namespace. This will become problematic when we'll try to execute the handler
in a container.

Let's retrieve their value from /proc intead so the values will always be in
the view of the user namespace the handler is running in and thus there wont be
any risk to leak such metadata to an unprivileged namespace.